code refactoring: removed the api package and moved classes into a different package added  custom docs in some classes
