+++ src/main/java/sainsbury/scraper/model/CummulatedProductResults.java
@@ -1,13 +1,11 @@
-package sainsbury.scrapper.model;
+package sainsbury.scraper.model;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-@EqualsAndHashCode
 @AllArgsConstructor
 @NoArgsConstructor
 public class CummulatedProductResults {
