+++ src/main/java/sainsbury/scraper/utils/JsoupDocumentRetriever.java
@@ -1,4 +1,4 @@
-package sainsbury.scrapper.utils;
+package sainsbury.scraper.utils;
 
 import org.apache.commons.lang3.StringUtils;
 import org.jsoup.nodes.Document;
@@ -10,6 +10,9 @@ import org.slf4j.LoggerFactory;
 import java.math.BigDecimal;
 import java.util.Optional;
 
+/**
+ * This class contains a collection of methods to extract information from a Document
+ */
 public final class JsoupDocumentRetriever {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(JsoupDocumentRetriever.class);
@@ -17,6 +20,11 @@ public final class JsoupDocumentRetriever {
     private JsoupDocumentRetriever() {
     }
 
+    /**
+     * This method selects all Elements from the Document
+     * @param document the grocery product page's document
+     * @return  Elements
+     */
     public static Elements getProductElements(Document document) {
         if (document != null) {
             return document.select(".productLister .gridItem");
@@ -24,6 +32,11 @@ public final class JsoupDocumentRetriever {
         return new Elements();
     }
 
+    /**
+     * This method selects the absolute link references to an element's information
+     * @param element the entity to extract relevant data from
+     * @return  absolute url
+     */
     public static String getProductSummaryUrl(Element element) {
         LOGGER.info("Getting product Summary Section Url....");
         return element.select("a")
@@ -31,6 +44,11 @@ public final class JsoupDocumentRetriever {
                 .attr("abs:href");
     }
 
+    /**
+     * This method extracts product's title
+     * @param document the grocery product page's document
+     * @return product title
+     */
     public static String getTitle(Document document) {
         LOGGER.info("Getting product title");
         return document.select(".productTitleDescriptionContainer > h1")
@@ -38,16 +56,25 @@ public final class JsoupDocumentRetriever {
     }
 
 
+    /**
+     * This method extracts product's price
+     * @param document the grocery product page's document
+     * @return products price
+     */
     public static BigDecimal getPrice(Document document) {
         LOGGER.info("Getting product Price");
         String rawPrice = document.select(".pricing > .pricePerUnit")
                 .first()
                 .text();
-        String cleanedPrice = StringUtils.replaceEach(rawPrice, new String[]{"£", "/unit"}, new String[]{"", ""});
-        return new BigDecimal(cleanedPrice);
+        String priceWithoutExtraText = StringUtils.replaceEach(rawPrice, new String[]{"£", "/unit"}, new String[]{"", ""});
+        return new BigDecimal(priceWithoutExtraText);
     }
 
-
+    /**
+     * This method extracts product's description
+     * @param document the grocery product page's document
+     * @return product's description
+     */
     public static String getDescription(Document document) {
         LOGGER.info("Getting product Description...");
         return document.select(".productText")
@@ -56,7 +83,11 @@ public final class JsoupDocumentRetriever {
 
     }
 
-
+    /**
+     * This method extracts product's energy
+     * @param document the grocery product page's document
+     * @return product's energy
+     */
     public static Integer getEnergy(Document document) {
         LOGGER.info("Getting product energy...");
         return Optional.ofNullable(document.select(".nutritionTable")
