--- src/main/java/sainsbury/scrapper/model/Product.java
@@ -1,44 +1,0 @@
-package sainsbury.scrapper.model;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Setter;
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-
-import java.math.BigDecimal;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-
-/**
- * This class is a representation of a single grocery product
- */
-public class Product {
-
-    private  String title;
-    private  Integer energy;
-    private  BigDecimal price;
-    private  String description;
-
-
-    @Override
-    public boolean equals(Object obj) {
-        return EqualsBuilder.reflectionEquals(this, obj);
-    }
-
-    @Override
-    public int hashCode() {
-        return HashCodeBuilder.reflectionHashCode(this);
-    }
-
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this);
-    }
-}
-
