+++ src/main/resources/log4j2.xml
@@ -9,7 +9,7 @@
                      filepattern="${logPath}/%d{YYYYMMddHHmmss}-fargo.log">
             <PatternLayout pattern="%d{YYYY-MM-dd HH:mm:ss} [%t] %-5p %c{1}:%L - %msg%n" />
             <Policies>
-                <SizeBasedTriggeringPolicy size="10 MB" />
+                <SizeBasedTriggeringPolicy size="5 MB" />
             </Policies>
             <DefaultRolloverStrategy max="20" />
         </RollingFile>
